Refactored all the reports components to make them more generic
