+++ modern/src/reports/ReportLayoutPage.js
@@ -18,7 +18,7 @@ const useStyles = makeStyles(theme => ({
   },
 }));
 
-const ReportLayoutPage = ({ reportFilterForm:ReportFilterForm, setItems, type, setType, children }) => {
+const ReportLayoutPage = ({ children, filter }) => {
   const classes = useStyles();
   return (
     <div className={classes.root}>
@@ -27,7 +27,7 @@ const ReportLayoutPage = ({ reportFilterForm:ReportFilterForm, setItems, type, s
         <Grid container spacing={2}>
           <Grid item xs={12} md={3} lg={2}>
             <Paper className={classes.form}>
-              <ReportFilterForm setItems={ setItems } type={ type } setType={ setType } />
+              {filter}
             </Paper>
           </Grid>
           <Grid item xs={12} md={9} lg={10}>
